Fix the incorrect typecast in vgt_host_bridge_cap_init
This commit is to fix the qemu segfault issue described in bug584.
The segfault happens in vgt_host_bridge_cap_init(), which incorrectly
converted dev to a vgt_vga_state_t pointer. However, here in this
routine, the device refereneced by dev is of type PCII440FXState.
Normally, although &o->host_dev is incorrectly used in this routine,
its value still falls into a vma of qemu, with its contents being zero.
Yet in some scenario, the value of &o->host_dev falls outside the vma
of qemu, and the segfault occurs.

Signed-off-by: Yu Zhang <yu.c.zhang@intel.com>